pyproject.toml
[project]
name = "secretaria"
version = "0.1.0"
description = "Asistente de correo que conecta vía IMAP a cualquier servidor y crea eventos en Google Calendar usando LLM"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.0.0",
    "requests>=2.25.0",
    "python-dateutil>=2.8.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
